BACKOFFICE: implement changes agreed on TCK meeting 2019-06-07
ref: https://github.com/dmn-tck/tck/pull/238#issuecomment-497744315

== Decision Service tests
There are a number of test cases which relies on Decision Service, and
which are not consistent with the other proposed PRs we currently have
open to agree on a basis for Decision Service testing from the TCK
API/infra point of view.

ACTION My proposal is to defer this tests at least until said PRs are
merged, and we have at least the basics of Decision Service we agree on.
This applies to test cases named with prefix decisionService_.
I can do by backoffice if we agree.

FINAL RESULT aligned test to TCK XSD/API/infra for Decision Service

== typeRef inconsistencies
There are a number of test cases which inconsistently have the typeRef
not on the variable element but the decision logic element. I would
prefer to have it all on the variable element

ACTION My proposal is to update tests to have the typeRef on the
variable element.
This applies to test cases named with prefix literal_ and invoke_ if I
checked correctly.
I can do by backoffice if we agree.

FINAL RESULT done