BACKOFFICE: implement changes agreed on TCK meeting 2019-06-07
ref: https://github.com/dmn-tck/tck/pull/238#issuecomment-497744315

== Expected result is a wrapped element
A number of tests cases expect the result to auto-wrap as a singleton
list.
For instance a Decision results in 123, while because variable typeRef
is a list of number, expected result is [123].
I am not sure where DMNv1.2 specify this.
I am open to be shown where this is mandated by the spec.

ACTION I cannot understand the result of these cases, and to me unless
shown otherwise on the DMNv1.2 spec, I propose to erase them.
This applies to test cases

decision_006
decision_bkm_004
invoke_003
literal_003
EDIT: also

decisionService_003
decisionService_003_a
I can do by backoffice if we agree.

FINAL RESULT commented as agreed on TCK meeting 2019-06-07